create object that maps the asset type to the title
