+++ graphql/types.ts
@@ -47,17 +47,6 @@ export enum AssetTypeEnum {
   Link = 'LINK'
 }
 
-/** Enum for the asset type icons. */
-export enum AssetTypeIconsEnum {
-  AndroidFile = 'mdi-android',
-  AndroidStore = 'mdi-google-play',
-  Domain = 'mdi-web',
-  File = 'mdi-file',
-  IosFile = 'mdi-apple-ios',
-  IosStore = 'mdi-apple',
-  Ip = 'mdi-network-outline',
-  Link = 'mdi-link-box-variant-outline'
-}
 
 /** Create asset mutation. */
 export type CreateAssetsMutation = {
